+++ benches/ball_tree.rs
@@ -12,7 +12,12 @@ fn build(c: &mut Criterion) {
     let array = ArrayView::from_shape((n, dim), &data).unwrap();
     c.bench_function("build", |b| {
         b.iter(|| {
-            BallTree::new(array, distance::EUCLIDEAN).expect("`array` is not empty");
+            BallTree::new(
+                array,
+                distance::euclidean_distance,
+                Some(distance::euclidean_reduced_distance),
+            )
+            .expect("`array` is not empty");
         })
     });
 }
@@ -24,7 +29,12 @@ fn query_radius(c: &mut Criterion) {
     let mut rng = StdRng::from_seed(*b"ball tree query_radius test seed");
     let data: Vec<f64> = (0..n * dim).map(|_| rng.gen()).collect();
     let array = ArrayView::from_shape((n, dim), &data).unwrap();
-    let tree = BallTree::new(array, distance::EUCLIDEAN).expect("`array` is not empty");
+    let tree = BallTree::new(
+        array,
+        distance::euclidean_distance,
+        Some(distance::euclidean_reduced_distance),
+    )
+    .expect("`array` is not empty");
     c.bench_function("query_radius", |b| {
         b.iter(|| {
             for i in 0..n {