+++ src/ball_tree.rs
@@ -1,45 +1,51 @@
-use crate::distance::Metric;
+use crate::distance::{self, Distance};
 use ndarray::{ArrayBase, ArrayView1, CowArray, Data, Ix2};
+use num_traits::{Float, FromPrimitive, Zero};
 use std::cmp;
 use std::collections::BinaryHeap;
 use std::convert::TryFrom;
-use std::error;
-use std::fmt;
 use std::mem::size_of;
-use std::ops::Range;
+use std::ops::{AddAssign, DivAssign, Range};
+use thiserror::Error;
 
 /// A data structure for nearest neighbor search in a multi-dimensional space.
-#[derive(Debug)]
-pub struct BallTree<'a, M>
+pub struct BallTree<'a, A>
 where
-    M: Metric,
+    A: Float,
 {
-    points: CowArray<'a, f64, Ix2>,
+    points: CowArray<'a, A, Ix2>,
     idx: Vec<usize>,
-    nodes: Vec<Node>,
-    metric: M,
+    nodes: Vec<Node<A>>,
+    distance: Distance<A>,
+    _reduced_distance: Distance<A>,
 }
 
-impl<'a, M> BallTree<'a, M>
+impl<'a, A> BallTree<'a, A>
 where
-    M: Metric,
+    A: Float + Zero + AddAssign + DivAssign + FromPrimitive,
 {
     /// Builds a ball tree using the given distance metric.
     ///
     /// # Errors
     ///
-    /// Returns an error if `points` is an empty array.
-    pub fn new<T>(points: T, metric: M) -> Result<Self, EmptyArrayError>
+    /// * `InvalidArray::Empty` if `points` is an empty array.
+    /// * `InvalidArray::NotContiguous` if any row in `points` is not
+    ///   contiguous in memory.
+    pub fn new<T>(
+        points: T,
+        distance: Distance<A>,
+        reduced_distance: Option<Distance<A>>,
+    ) -> Result<Self, InvalidArray>
     where
-        T: Into<CowArray<'a, f64, Ix2>>,
+        T: Into<CowArray<'a, A, Ix2>>,
     {
         let points = points.into();
-        let n_points: usize = *points
-            .shape()
-            .first()
-            .expect("ArrayView2 should have two dimensions");
+        let n_points: usize = points.nrows();
         if n_points == 0 {
-            return Err(EmptyArrayError {});
+            return Err(InvalidArray::Empty);
+        }
+        if !points.row(0).is_standard_layout() {
+            return Err(InvalidArray::NotContiguous);
         }
 
         let height = u32::try_from(size_of::<usize>() * 8).expect("smaller than u32::max_value()")
@@ -48,13 +54,39 @@ where
 
         let mut idx: Vec<usize> = (0..n_points).collect();
         let mut nodes = vec![Node::default(); size];
-        build_subtree(&mut nodes, &mut idx, &points, 0, 0..n_points, &metric);
+        let reduced_distance = reduced_distance.unwrap_or(distance);
+        build_subtree(
+            &mut nodes,
+            &mut idx,
+            &points,
+            0,
+            0..n_points,
+            distance,
+            reduced_distance,
+        );
         Ok(BallTree {
             points,
             idx,
             nodes,
-            metric,
+            distance,
+            _reduced_distance: reduced_distance,
         })
+    }
+
+    /// Builds a ball tree with a euclidean distance metric.
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if `points` is an empty array.
+    pub fn euclidean<T>(points: T) -> Result<Self, InvalidArray>
+    where
+        T: Into<CowArray<'a, A, Ix2>>,
+    {
+        Self::new(
+            points,
+            distance::euclidean,
+            Some(distance::euclidean_reduced),
+        )
     }
 
     /// Finds the nearest neighbor and its distance in the tree.
@@ -66,16 +98,13 @@ where
     /// use petal_neighbors::{BallTree, distance};
     ///
     /// let points = array![[1., 1.], [1., 2.], [9., 9.]];
-    /// let tree = BallTree::new(points, distance::EUCLIDEAN).expect("non-empty input");
+    /// let tree = BallTree::euclidean(points).expect("non-empty input");
     /// let (index, distance) = tree.query_nearest(&[8., 8.]);
     /// assert_eq!(index, 2);  // points[2] is the nearest.
     /// assert!((2_f64.sqrt() - distance).abs() < 1e-8);
     /// ```
-    pub fn query_nearest<'p, P>(&self, point: P) -> (usize, f64)
-    where
-        P: Copy + IntoIterator<Item = &'p f64>,
-    {
-        self.nearest_neighbor_in_subtree(point, 0, std::f64::INFINITY)
+    pub fn query_nearest(&self, point: &[A]) -> (usize, A) {
+        self.nearest_neighbor_in_subtree(point, 0, A::infinity())
             .expect("0 is a valid index")
     }
 
@@ -89,16 +118,13 @@ where
     /// use petal_neighbors::{BallTree, distance};
     ///
     /// let points = array![[1., 1.], [1., 2.], [9., 9.]];
-    /// let tree = BallTree::new(points, distance::EUCLIDEAN).expect("non-empty input");
+    /// let tree = BallTree::euclidean(points).expect("non-empty input");
     /// let (indices, distances) = tree.query(&[3., 3.], 2);
     /// assert_eq!(indices, &[1, 0]);  // points[1] is the nearest, followed by points[0].
     /// ```
-    pub fn query<'p, P>(&self, point: P, k: usize) -> (Vec<usize>, Vec<f64>)
-    where
-        P: Copy + IntoIterator<Item = &'p f64>,
-    {
+    pub fn query(&self, point: &[A], k: usize) -> (Vec<usize>, Vec<A>) {
         let mut neighbors = BinaryHeap::with_capacity(k);
-        self.nearest_k_neighbors_in_subtree(point, 0, std::f64::INFINITY, k, &mut neighbors);
+        self.nearest_k_neighbors_in_subtree(point, 0, A::infinity(), k, &mut neighbors);
         let sorted = neighbors.into_sorted_vec();
         let indices = sorted.iter().map(|v| v.idx).collect();
         let distances = sorted.iter().map(|v| v.distance).collect();
@@ -115,14 +141,11 @@ where
     /// use petal_neighbors::{BallTree, distance};
     ///
     /// let points = array![[1., 0.], [2., 0.], [9., 0.]];
-    /// let tree = BallTree::new(points, distance::EUCLIDEAN).expect("non-empty input");
+    /// let tree = BallTree::euclidean(points).expect("non-empty input");
     /// let indices = tree.query_radius(&[3., 0.], 1.5);
     /// assert_eq!(indices, &[1]);  // The distance to points[1] is less than 1.5.
     /// ```
-    pub fn query_radius<'p, P>(&self, point: P, distance: f64) -> Vec<usize>
-    where
-        P: Copy + IntoIterator<Item = &'p f64>,
-    {
+    pub fn query_radius(&self, point: &[A], distance: A) -> Vec<usize> {
         self.neighbors_within_radius_in_subtree(point, distance, 0)
     }
 
@@ -131,26 +154,27 @@ where
     /// # Panics
     ///
     /// Panics if `root` is out of bound.
-    fn nearest_neighbor_in_subtree<'p, P>(
+    fn nearest_neighbor_in_subtree(
         &self,
-        point: P,
+        point: &[A],
         root: usize,
-        radius: f64,
-    ) -> Option<(usize, f64)>
-    where
-        P: Copy + IntoIterator<Item = &'p f64>,
-    {
+        radius: A,
+    ) -> Option<(usize, A)> {
         let root_node = &self.nodes[root];
-        let lower_bound = self.nodes[root].distance_lower_bound(point, &self.metric);
+        let lower_bound = self.nodes[root].distance_lower_bound(point, self.distance);
         if lower_bound > radius {
             return None;
         }
 
         if root_node.is_leaf {
             let (min_i, min_dist) = self.idx[root_node.range.clone()].iter().fold(
-                (0, std::f64::INFINITY),
+                (0, A::infinity()),
                 |(min_i, min_dist), &i| {
-                    let dist = self.metric.distance(point, self.points.row(i));
+                    let distance = self.distance;
+                    let dist = distance(
+                        point,
+                        self.points.row(i).as_slice().expect("standard layout"),
+                    );
 
                     if dist < min_dist {
                         (i, dist)
@@ -167,8 +191,8 @@ where
         } else {
             let child1 = root * 2 + 1;
             let child2 = child1 + 1;
-            let lb1 = self.nodes[child1].distance_lower_bound(point, &self.metric);
-            let lb2 = self.nodes[child2].distance_lower_bound(point, &self.metric);
+            let lb1 = self.nodes[child1].distance_lower_bound(point, self.distance);
+            let lb2 = self.nodes[child2].distance_lower_bound(point, self.distance);
             let (child1, child2) = if lb1 < lb2 {
                 (child1, child2)
             } else {
@@ -194,18 +218,16 @@ where
     /// # Panics
     ///
     /// Panics if `root` is out of bound.
-    fn nearest_k_neighbors_in_subtree<'p, P>(
+    fn nearest_k_neighbors_in_subtree(
         &self,
-        point: P,
+        point: &[A],
         root: usize,
-        radius: f64,
+        radius: A,
         k: usize,
-        neighbors: &mut BinaryHeap<Neighbor>,
-    ) where
-        P: Copy + IntoIterator<Item = &'p f64>,
-    {
+        neighbors: &mut BinaryHeap<Neighbor<A>>,
+    ) {
         let root_node = &self.nodes[root];
-        if root_node.distance_lower_bound(point, &self.metric) > radius {
+        if root_node.distance_lower_bound(point, self.distance) > radius {
             return;
         }
 
@@ -213,7 +235,11 @@ where
             self.idx[root_node.range.clone()]
                 .iter()
                 .filter_map(|&i| {
-                    let dist = self.metric.distance(point, self.points.row(i));
+                    let distance = self.distance;
+                    let dist = distance(
+                        point,
+                        self.points.row(i).as_slice().expect("standard layout"),
+                    );
 
                     if dist < radius {
                         Some(Neighbor::new(i, dist))
@@ -233,8 +259,8 @@ where
         } else {
             let child1 = root * 2 + 1;
             let child2 = child1 + 1;
-            let lb1 = self.nodes[child1].distance_lower_bound(point, &self.metric);
-            let lb2 = self.nodes[child2].distance_lower_bound(point, &self.metric);
+            let lb1 = self.nodes[child1].distance_lower_bound(point, self.distance);
+            let lb2 = self.nodes[child2].distance_lower_bound(point, self.distance);
             let (child1, child2) = if lb1 < lb2 {
                 (child1, child2)
             } else {
@@ -250,22 +276,19 @@ where
     /// # Panics
     ///
     /// Panics if `root` is out of bound.
-    fn neighbors_within_radius_in_subtree<'p, P>(
+    fn neighbors_within_radius_in_subtree(
         &self,
-        point: P,
-        radius: f64,
+        point: &[A],
+        radius: A,
         root: usize,
-    ) -> Vec<usize>
-    where
-        P: Copy + IntoIterator<Item = &'p f64>,
-    {
+    ) -> Vec<usize> {
         let mut neighbors = Vec::new();
         let mut subtrees_to_visit = vec![root];
 
         loop {
             let subroot = subtrees_to_visit.pop().expect("should not be empty");
             let root_node = &self.nodes[subroot];
-            let (lb, ub) = root_node.distance_bounds(point, &self.metric);
+            let (lb, ub) = root_node.distance_bounds(point, self.distance);
 
             if lb > radius {
                 if subtrees_to_visit.is_empty() {
@@ -279,7 +302,11 @@ where
                 neighbors.extend(self.idx[root_node.range.clone()].iter().cloned());
             } else if root_node.is_leaf {
                 neighbors.extend(self.idx[root_node.range.clone()].iter().filter_map(|&i| {
-                    let dist = self.metric.distance(point, self.points.row(i));
+                    let distance = self.distance;
+                    let dist = distance(
+                        point,
+                        self.points.row(i).as_slice().expect("standard layout"),
+                    );
                     if dist < radius {
                         Some(i)
                     } else {
@@ -300,127 +327,137 @@ where
     }
 }
 
-/// An error returned when an empty array is given as an argument to a function
-/// requiring a non-empty array.
-#[derive(Debug)]
-pub struct EmptyArrayError;
-
-impl fmt::Display for EmptyArrayError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "input array is empty")
-    }
+/// An error returned when an array is not suitable to build a `BallTree`.
+#[derive(Debug, Error)]
+pub enum InvalidArray {
+    #[error("array is empty")]
+    Empty,
+    #[error("array is not contiguous in memory")]
+    NotContiguous,
 }
-
-impl error::Error for EmptyArrayError {}
 
 #[derive(Clone, Debug)]
-struct Neighbor {
+struct Neighbor<A>
+where
+    A: Float,
+{
     pub idx: usize,
-    pub distance: f64,
+    pub distance: A,
 }
 
-impl Neighbor {
+impl<A> Neighbor<A>
+where
+    A: Float,
+{
     #[must_use]
-    pub fn new(idx: usize, distance: f64) -> Self {
+    pub fn new(idx: usize, distance: A) -> Self {
         Self { idx, distance }
     }
 }
 
-impl Ord for Neighbor {
+impl<A> Ord for Neighbor<A>
+where
+    A: Float,
+{
     #[must_use]
     fn cmp(&self, other: &Self) -> cmp::Ordering {
         self.partial_cmp(other).unwrap()
     }
 }
 
-impl PartialOrd for Neighbor {
+impl<A> PartialOrd for Neighbor<A>
+where
+    A: Float,
+{
     #[must_use]
     fn partial_cmp(&self, other: &Self) -> Option<cmp::Ordering> {
         self.distance.partial_cmp(&other.distance)
     }
 }
 
-impl PartialEq for Neighbor {
+impl<A> PartialEq for Neighbor<A>
+where
+    A: Float,
+{
     #[must_use]
     fn eq(&self, other: &Self) -> bool {
         self.distance == other.distance
     }
 }
 
-impl Eq for Neighbor {}
+impl<A> Eq for Neighbor<A> where A: Float {}
 
 /// A node containing a range of points in a ball tree.
 #[derive(Clone, Debug)]
-struct Node {
+struct Node<A> {
     range: Range<usize>,
-    centroid: Vec<f64>,
-    radius: f64,
+    centroid: Vec<A>,
+    radius: A,
     is_leaf: bool,
 }
 
-impl Node {
+impl<A> Node<A>
+where
+    A: Float + Zero + AddAssign + DivAssign + FromPrimitive,
+{
     /// Computes the centroid of the node.
+    ///
+    /// # Panics
+    ///
+    /// Panics if any row in `points` is not contiguous in memory.
     #[allow(clippy::cast_precision_loss)] // The precision provided by 54-bit-wide mantissa is
                                           // good enough in computing mean.
-    fn init<M>(&mut self, points: &CowArray<f64, Ix2>, idx: &[usize], metric: &M)
-    where
-        M: Metric,
-    {
-        let mut sum = idx.iter().fold(vec![0_f64; points.ncols()], |mut sum, &i| {
-            for (s, v) in sum.iter_mut().zip(points.row(i)) {
-                *s += *v;
-            }
-            sum
-        });
-        sum.iter_mut().for_each(|v| *v /= idx.len() as f64);
+    fn init(&mut self, points: &CowArray<A, Ix2>, idx: &[usize], distance: Distance<A>) {
+        let mut sum = idx
+            .iter()
+            .fold(vec![A::zero(); points.ncols()], |mut sum, &i| {
+                for (s, v) in sum.iter_mut().zip(points.row(i)) {
+                    *s += *v;
+                }
+                sum
+            });
+        let len = A::from_usize(idx.len()).expect("approximation");
+        sum.iter_mut().for_each(|v| *v /= len);
         self.centroid = sum;
 
-        self.radius = idx.iter().fold(0., |max, &i| {
-            f64::max(
-                metric.distance(
-                    &self.centroid,
-                    points.row(i).to_slice().expect("should be contiguous"),
-                ),
+        self.radius = idx.iter().fold(A::zero(), |max, &i| {
+            A::max(
+                distance(&self.centroid, &points.row(i).as_slice().unwrap()),
                 max,
             )
         });
     }
 
-    fn distance_bounds<'p, P, M>(&self, point: P, metric: &M) -> (f64, f64)
-    where
-        P: IntoIterator<Item = &'p f64>,
-        M: Metric,
-    {
-        let centroid_dist = metric.distance(point, &self.centroid);
+    fn distance_bounds(&self, point: &[A], distance: Distance<A>) -> (A, A) {
+        let centroid_dist = distance(point, &self.centroid);
         let mut lb = centroid_dist - self.radius;
-        if lb < 0. {
-            lb = 0.;
+        if lb < A::zero() {
+            lb = A::zero();
         }
         let ub = centroid_dist + self.radius;
         (lb, ub)
     }
 
-    fn distance_lower_bound<'p, P, M>(&self, point: P, metric: &M) -> f64
-    where
-        P: IntoIterator<Item = &'p f64>,
-        M: Metric,
-    {
-        let centroid_dist = metric.distance(point, &self.centroid);
+    fn distance_lower_bound(&self, point: &[A], distance: Distance<A>) -> A {
+        let centroid_dist = distance(point, &self.centroid);
         let lb = centroid_dist - self.radius;
-        if lb < 0. {
-            0.
+        if lb < A::zero() {
+            A::zero()
         } else {
             lb
         }
     }
 }
 
-impl Default for Node {
+impl<A> Default for Node<A>
+where
+    A: Float + Zero,
+{
     fn default() -> Self {
         Self {
             range: (0..0),
             centroid: Vec::new(),
-            radius: 0.,
+            radius: A::zero(),
             is_leaf: false,
         }
     }
@@ -431,22 +468,23 @@ impl Default for Node {
 /// # Panics
 ///
 /// Panics if `root` or `range` is out of range.
-fn build_subtree<M>(
-    nodes: &mut [Node],
+fn build_subtree<A>(
+    nodes: &mut [Node<A>],
     idx: &mut [usize],
-    points: &CowArray<f64, Ix2>,
+    points: &CowArray<A, Ix2>,
     root: usize,
     range: Range<usize>,
-    metric: &M,
+    distance: Distance<A>,
+    reduced_distance: Distance<A>,
 ) where
-    M: Metric,
+    A: Float + AddAssign + DivAssign + FromPrimitive,
 {
     let n_nodes = nodes.len();
     let mut root_node = nodes.get_mut(root).expect("root node index out of range");
     root_node.init(
         points,
         &idx.get(range.clone()).expect("invalid subtree range"),
-        metric,
+        distance,
     );
     root_node.range = range.clone();
     let left = root * 2 + 1;
@@ -462,8 +500,24 @@ fn build_subtree<M>(
     halve_node_indices(&mut idx[range.clone()], &col);
 
     let mid = (range.start + range.end) / 2;
-    build_subtree(nodes, idx, points, left, range.start..mid, metric);
-    build_subtree(nodes, idx, points, left + 1, mid..range.end, metric);
+    build_subtree(
+        nodes,
+        idx,
+        points,
+        left,
+        range.start..mid,
+        distance,
+        reduced_distance,
+    );
+    build_subtree(
+        nodes,
+        idx,
+        points,
+        left + 1,
+        mid..range.end,
+        distance,
+        reduced_distance,
+    );
 }
 
 /// Divides the node index array into two equal-sized parts.
@@ -471,7 +525,10 @@ fn build_subtree<M>(
 /// # Panics
 ///
 /// Panics if `col` is empty.
-fn halve_node_indices(idx: &mut [usize], col: &ArrayView1<f64>) {
+fn halve_node_indices<A>(idx: &mut [usize], col: &ArrayView1<A>)
+where
+    A: Float,
+{
     let (mut first, mut last) = (0, idx.len() - 1);
     let mid = idx.len() / 2;
     loop {
@@ -500,9 +557,10 @@ fn halve_node_indices(idx: &mut [usize], col: &ArrayView1<f64>) {
 ///
 /// Panics if either `matrix` or `idx` is empty, or any element of `idx` is
 /// greater than or equal to the number of rows in `matrix`.
-fn max_spread_column<D>(matrix: &ArrayBase<D, Ix2>, idx: &[usize]) -> usize
+fn max_spread_column<A, S>(matrix: &ArrayBase<S, Ix2>, idx: &[usize]) -> usize
 where
-    D: Data<Elem = f64>,
+    A: Float,
+    S: Data<Elem = A>,
 {
     let mut spread_iter = matrix
         .gencolumns()
@@ -512,7 +570,7 @@ where
                 .iter()
                 .skip(1)
                 .fold((col[idx[0]], col[idx[0]]), |(min, max), &i| {
-                    (f64::min(min, col[i]), f64::max(max, col[i]))
+                    (A::min(min, col[i]), A::max(max, col[i]))
                 });
             max - min
         })
@@ -547,8 +605,12 @@ mod test {
     #[should_panic]
     fn ball_tree_empty() {
         let data: [[f64; 0]; 0] = [];
-        let tree =
-            BallTree::new(aview2(&data), distance::EUCLIDEAN).expect("`data` should not be empty");
+        let tree = BallTree::new(
+            aview2(&data),
+            distance::euclidean,
+            Some(distance::euclidean_reduced),
+        )
+        .expect("`data` should not be empty");
         let point = [0., 0.];
         tree.query_nearest(&point);
     }
@@ -556,7 +618,12 @@ mod test {
     #[test]
     fn ball_tree_3() {
         let array = array![[1., 1.], [1., 1.1], [9., 9.]];
-        let tree = BallTree::new(array, distance::EUCLIDEAN).expect("`array` should not be empty");
+        let tree = BallTree::new(
+            array,
+            distance::euclidean,
+            Some(distance::euclidean_reduced),
+        )
+        .expect("`array` should not be empty");
 
         let point = [0., 0.];
         let neighbor = tree.query_nearest(&point);
@@ -605,7 +672,12 @@ mod test {
             [-2.0, 3.0],
             [-2.2, 3.1],
         ];
-        let tree = BallTree::new(array, distance::EUCLIDEAN).expect("`array` should not be empty");
+        let tree = BallTree::new(
+            array,
+            distance::euclidean,
+            Some(distance::euclidean_reduced),
+        )
+        .expect("`array` should not be empty");
 
         let point = [1., 2.];
         let neighbor = tree.query_nearest(&point);
@@ -625,7 +697,12 @@ mod test {
             [1.0, 1.0],
             [1.0, 1.0],
         ];
-        let tree = BallTree::new(array, distance::EUCLIDEAN).expect("`array` should not be empty");
+        let tree = BallTree::new(
+            array,
+            distance::euclidean,
+            Some(distance::euclidean_reduced),
+        )
+        .expect("`array` should not be empty");
 
         let point = [1., 2.];
         let neighbor = tree.query_nearest(&point);
@@ -637,16 +714,17 @@ mod test {
         const DIMENSION: usize = 3;
 
         let array = Array::random((40, DIMENSION), Uniform::new(0., 1.));
-        let bt =
-            BallTree::new(array.view(), distance::EUCLIDEAN).expect("`array` should not be empty");
+        let bt = BallTree::new(array.view(), distance::euclidean, None)
+            .expect("`array` should not be empty");
         for _ in 0..10 {
             let query = Array::random(DIMENSION, Uniform::new(0., 1.));
-            let (_, bt_distances) = bt.query(query.as_slice().expect("should be contiguous"), 5);
+            let (_, bt_distances) =
+                bt.query(query.as_slice().expect("should be contiguous in memory"), 5);
             let naive_neighbors = naive_k_nearest_neighbors(
                 &array,
-                query.as_slice().expect("should be contiguous"),
+                query.as_slice().expect("should be contiguous in memory"),
                 5,
-                distance::EUCLIDEAN,
+                distance::euclidean,
             );
             for (bt_dist, naive_neighbor) in bt_distances.iter().zip(naive_neighbors.iter()) {
                 assert!(approx::abs_diff_eq!(*bt_dist, naive_neighbor.distance));
@@ -657,7 +735,12 @@ mod test {
     #[test]
     fn ball_tree_query_radius() {
         let array = array![[0.], [2.], [3.], [4.], [6.], [8.], [10.]];
-        let bt = BallTree::new(array, distance::EUCLIDEAN).expect("`array` should not be empty");
+        let bt = BallTree::new(
+            array,
+            distance::euclidean,
+            Some(distance::euclidean_reduced),
+        )
+        .expect("`array` should not be empty");
 
         let neighbors = bt.query_radius(&[0.1], 1.);
         assert_eq!(neighbors, &[0]);
@@ -675,12 +758,12 @@ mod test {
         let array = array![[0., 1.], [0., 9.], [0., 2.]];
         let idx: [usize; 3] = [0, 1, 2];
         let mut node = Node::default();
-        node.init(&array.view().into(), &idx, &distance::EUCLIDEAN);
+        node.init(&array.view().into(), &idx, distance::euclidean);
         assert_eq!(node.centroid, [0., 4.]);
         assert_eq!(node.radius, 5.);
 
         let idx: [usize; 2] = [0, 2];
-        node.init(&array.into(), &idx, &distance::EUCLIDEAN);
+        node.init(&array.into(), &idx, distance::euclidean);
         assert_eq!(node.centroid, [0., 1.5]);
     }
 
@@ -752,24 +835,29 @@ mod test {
         assert_eq!(super::max_spread_column(&aview2(&data), &idx), 1);
     }
 
-    fn naive_k_nearest_neighbors<'a, D, M>(
-        neighbors: &'a ArrayBase<D, Ix2>,
-        point: &[f64],
+    /// Finds k nearest neighbors by compuing distance to every point.
+    ///
+    /// # Panics
+    ///
+    /// Panics if any row in `neighbors` is not contiguous in memory.
+    fn naive_k_nearest_neighbors<'a, A, S>(
+        neighbors: &'a ArrayBase<S, Ix2>,
+        point: &[A],
         k: usize,
-        metric: M,
-    ) -> Vec<Neighbor>
+        distance: Distance<A>,
+    ) -> Vec<Neighbor<A>>
     where
-        D: Data<Elem = f64>,
-        M: Metric,
+        A: Float,
+        S: Data<Elem = A>,
     {
         let mut knn = neighbors
             .axis_iter(Axis(0))
             .enumerate()
             .map(|(i, n)| Neighbor {
                 idx: i,
-                distance: metric.distance(n.to_slice().unwrap(), point),
+                distance: distance(n.to_slice().unwrap(), point),
             })
-            .collect::<Vec<Neighbor>>();
+            .collect::<Vec<Neighbor<A>>>();
         knn.sort();
         knn[0..k].to_vec()
     }